after cleanup, pending propagating root
