--- data-access/src/app/domain/contexts/cases/service-ticket/v1/sync-domain-events/handlers/index.ts
@@ -1,9 +1,0 @@
-// import { EventHandlerTuple } from "../../../../../../../../../seedwork/domain-seedwork/aggregate-root";
-// import { ServiceTicketV1CreatedHandlerAddActivityDetailRecordTuple } from "./service-ticket-v1-created--add-activity-detail-record";
-
-// export const ServiceTicketV1SyncDomainEventHandlers: EventHandlerTuple<any, any>[] = [
-//     ServiceTicketV1CreatedHandlerAddActivityDetailRecordTuple
-// ]
-
-
-
