+++ data-access/src/app/domain/contexts/cases/service-ticket/v1/sync-domain-events/service-ticket-v1.sync-domain-event-handlers.ts
@@ -1,4 +1,4 @@
-import { serviceTicketV1CreatedSyncDomainEventHandler } from "./service-ticket-v1-created.sync-domain-event";
+import { serviceTicketV1CreatedSyncDomainEventHandler } from "./types/service-ticket-v1-created.sync-domain-event";
 
 export const ServiceTicketV1SyncDomainEventHandlers= {
     ServiceTicketV1CreatedSyncDomainEvent: serviceTicketV1CreatedSyncDomainEventHandler,
