--- data-access/src/app/domain/contexts/cases/service-ticket/v1/sync-domain-events/types/service-ticket-v1-created.ts
@@ -1,10 +1,0 @@
-// import { CustomDomainEventImpl } from "../../../../../../../../../seedwork/domain-seedwork/domain-event"
-// import { MemberEntityReference } from "../../../../../community/member/member"
-
-// export interface ServiceTicketV1CreatedSyncDomainEventProps {
-//     requestor: MemberEntityReference
-// }
-// export class ServiceTicketV1CreatedSyncDomainEvent extends CustomDomainEventImpl<ServiceTicketV1CreatedSyncDomainEventProps> {}
-
-
-
